+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 
 These versions:
 
+- Bump to OS26.
 - Remove deprecated types.
 - Remove subscription feature.